requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django_rest_knox==4.1.0
django_rest_passwordreset==1.1.0
Faker==8.2.1
